adding support for drawing text
Unrolled for loop in Set_Pixel to save 2uSec.
Seem to be able to use significantly less than 10uSec in bit pattern to lights. But 30uSec at end of frame seems to be a tight minimum. Need to perform more testing to discover minimum timing.
Current version of Put_Char is time consuming because of the 40 calls to Set_Pixel.